move the logic for writing the avro files under the 'manifest_file' and 'manifest_list' namespaces, so we can use the constants for the field ids
