+++ src/include/metadata/iceberg_manifest.hpp
@@ -5,7 +5,13 @@
 #include "duckdb/common/vector.hpp"
 #include "duckdb/common/types/value.hpp"
 
+#include "duckdb/function/copy_function.hpp"
+#include "duckdb/execution/execution_context.hpp"
+#include "duckdb/parallel/thread_context.hpp"
+
 namespace duckdb {
+
+struct IcebergTableInformation;
 
 using sequence_number_t = int64_t;
 
@@ -82,8 +88,61 @@ struct IcebergManifestFile {
 	}
 
 public:
+public:
 	string path;
 	vector<IcebergManifestEntry> data_files;
 };
+
+namespace manifest_file {
+
+static constexpr const int32_t STATUS = 0;
+static constexpr const int32_t SNAPSHOT_ID = 1;
+static constexpr const int32_t SEQUENCE_NUMBER = 3;
+static constexpr const int32_t FILE_SEQUENCE_NUMBER = 4;
+static constexpr const int32_t DATA_FILE = 2;
+static constexpr const int32_t CONTENT = 134;
+static constexpr const int32_t FILE_PATH = 100;
+static constexpr const int32_t FILE_FORMAT = 101;
+static constexpr const int32_t PARTITION = 102;
+static constexpr const int32_t RECORD_COUNT = 103;
+static constexpr const int32_t FILE_SIZE_IN_BYTES = 104;
+// static constexpr const int32_t BLOCK_SIZE_IN_BYTES = 105; // (deprecated)
+// static constexpr const int32_t FILE_ORDINAL = 106; // (deprecated)
+// static constexpr const int32_t SORT_COLUMNS = 107; // (deprecated)
+// static constexpr const int32_t SORT_COLUMNS_ELEMENT = 112; // (deprecated)
+static constexpr const int32_t COLUMN_SIZES = 108;
+static constexpr const int32_t COLUMN_SIZES_KEY = 117;
+static constexpr const int32_t COLUMN_SIZES_VALUE = 118;
+static constexpr const int32_t VALUE_COUNTS = 109;
+static constexpr const int32_t VALUE_COUNTS_KEY = 119;
+static constexpr const int32_t VALUE_COUNTS_VALUE = 120;
+static constexpr const int32_t NULL_VALUE_COUNTS = 110;
+static constexpr const int32_t NULL_VALUE_COUNTS_KEY = 121;
+static constexpr const int32_t NULL_VALUE_COUNTS_VALUE = 122;
+static constexpr const int32_t NAN_VALUE_COUNTS = 137;
+static constexpr const int32_t NAN_VALUE_COUNTS_KEY = 138;
+static constexpr const int32_t NAN_VALUE_COUNTS_VALUE = 139;
+// static constexpr const int32_t DISTINCT_COUNTS = 111; // (deprecated)
+static constexpr const int32_t LOWER_BOUNDS = 125;
+static constexpr const int32_t LOWER_BOUNDS_KEY = 126;
+static constexpr const int32_t LOWER_BOUNDS_VALUE = 127;
+static constexpr const int32_t UPPER_BOUNDS = 128;
+static constexpr const int32_t UPPER_BOUNDS_KEY = 129;
+static constexpr const int32_t UPPER_BOUNDS_VALUE = 130;
+// static constexpr const int32_t KEY_METADATA = 131; // (optional)
+static constexpr const int32_t SPLIT_OFFSETS = 132;
+static constexpr const int32_t SPLIT_OFFSETS_ELEMENT = 133;
+static constexpr const int32_t EQUALITY_IDS = 135;
+static constexpr const int32_t EQUALITY_IDS_ELEMENT = 136;
+static constexpr const int32_t SORT_ORDER_ID = 140;
+static constexpr const int32_t FIRST_ROW_ID = 142;
+static constexpr const int32_t REFERENCED_DATA_FILE = 143;
+static constexpr const int32_t CONTENT_OFFSET = 144;
+static constexpr const int32_t CONTENT_SIZE_IN_BYTES = 145;
+
+idx_t WriteToFile(IcebergTableInformation &table_info, const IcebergManifestFile &manifest_file,
+                  CopyFunction &copy_function, DatabaseInstance &db, ClientContext &context);
+
+} // namespace manifest_file
 
 } // namespace duckdb
